requirements.txt
gradio==4.17.0
fastapi==0.110.0
uvicorn==0.27.1
openai==1.12.0
pymupdf==1.23.21
pillow==9.5.0
astrapy==0.4.1
numpy==1.26.4